backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.0
langchain==0.0.335
faiss-cpu==1.7.4
pydantic==2.4.2
python-dotenv==1.0.0
transformers==4.35.0
torch>=2.6.0
sentence-transformers==2.2.2
huggingface-hub==0.17.3
google-generativeai==0.7.2
python-multipart==0.0.6
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
email-validator==2.1.0.post1